Refactor and lots of comments added
- separated things into their own files
- you can now encrypt something for the past
