--- src/age/chunked.ts
@@ -1,25 +1,0 @@
-/*
-    e.g. chunked("hello world", 2, ".") returns
-    ["he.", "ll.", "o .", "wo.", "rl.", "d."]
- */
-function chunked(input: string, chunkSize: number, suffix = ""): Array<string> {
-    const output = []
-    let currentChunk = ""
-    for (let i = 0, chunks = 0; i < input.length; i++) {
-        currentChunk += input[i]
-
-        const posInChunk = i - (chunks * chunkSize)
-
-        if (posInChunk === chunkSize - 1) {
-            output.push(currentChunk + suffix)
-            currentChunk = ""
-            chunks++
-        } else if (i === input.length - 1) {
-            output.push(currentChunk + suffix)
-        }
-    }
-
-    return output
-}
-
-export {chunked}
