+++ src/schema/webform-schema.ts
@@ -5,7 +5,6 @@ const webFormSchema = yup.object({
     ciphertext: yup.string().nullable(true).optional(),
     decryptionTime: yup.number()
         .positive()
-        // .moreThan(Date.now(), "Decryption time must be in the future!")
         .required()
 }).required()
