+++ test/age/compatibility.test.ts
@@ -1,27 +1,25 @@
 import {test} from "mocha"
 import {expect} from "chai"
 import {timelockDecrypt} from "../../src/drand/timelock"
-import {MockDrandClient} from "./mock-drand-client"
+import {MockDrandClient} from "../drand/mock-drand-client"
 
-// needs reenabled once the `toField` impl in go is corrected
-test.skip("payloads encrypted with the go impl should decrypt successfully", async () => {
+test("payloads encrypted with the go impl should decrypt successfully", async () => {
     const validBeacon = {
-        round: 2627830,
-        randomness: "d6d843144ce19f63318048f2715cc17298ae6f2dfe982705c2e41a9f63e63b04",
-        signature: "8f1385df2916a38f6146bdbc269d7492398621ba005e9188e7854bfd7269ac09299474d03d31e85a699f8b050a88805019502a5ea96762752207878753827c1ac51e6c1bacec96bff495d696226c6b485b68e076ac760b883732e3fd457ba197"
+        round: 20000,
+        randomness: "49891540810f410d114b25bdd47e6c14d316da3e513a82e91cc705a9bbb43fb3",
+        signature: "91130cdcb1d7cbc402bbb10caecf3eaa2aa041a8bdc5a3791f24c033431bc0c7a0275c2b76fdcc2031dd613b2641a1a715eb6e97dce0a9c0325426315d5da276bd5ebbf07d2e6a082697e31db1d3e8f7813797d3c5a7ca95e90167404426a7c3"
     }
     const mockClient = new MockDrandClient(validBeacon)
-    const payloadFromGoImpl =  "-----BEGIN AGE ENCRYPTED FILE-----\n"+
-    "YWdlLWVuY3J5cHRpb24ub3JnL3YxCi0+IHRsb2NrIDI2Mjc4MzAgNzY3Mjc5N2Y1\n"+
-    "NDhmM2Y0NzQ4YWM0YmYzMzUyZmM2YzZiNjQ2OGM5YWQ0MGFkNDU2YTM5NzU0NWM2\n"+
-    "ZTJkZjViZgpncm1LcHZxdm1WL1FRNlhFbzNqbDlPWlZFb01OSTM3dCtZN3Y4MWV6\n"+
-    "NDZpVS9QTnArOTVoNEFnamltclFBYnlCCmtQdmwvZGZYdmtYVHNNbmYrcnhPRVQz\n"+
-    "WVpFdlhCeXVLNU1mVlB0VFBwWW8KLS0tIEdnV3dQSllaTFpPZEtpZDBPVWxVZXAv\n"+
-    "YnpzL3k0d1VaV1JkZzc3bytLVGcKSsSWLbyYFCCbm6NSBPD/wwhFIFWww5+BjsEN\n"+
-    "CIokjx+3bwBtmvqMXpKgB4U=\n"+
-    "-----END AGE ENCRYPTED FILE-----"
-
+    const payloadFromGoImpl = "-----BEGIN AGE ENCRYPTED FILE-----\n" +
+        "YWdlLWVuY3J5cHRpb24ub3JnL3YxCi0+IHRsb2NrIDIwMDAwIDc2NzI3OTdmNTQ4\n" +
+        "ZjNmNDc0OGFjNGJmMzM1MmZjNmM2YjY0NjhjOWFkNDBhZDQ1NmEzOTc1NDVjNmUy\n" +
+        "ZGY1YmYKc3lqZWdrbGdmZlRaQ3hITEN1YUlZakhDRGdZNTBUK0owNk1Zd1dMdjZB\n" +
+        "YUExN1o2YzVWWnJ5cTFBNWNLOWl4cQpUd2I5S3p0Z1NxRjRhdlNRbXRJN0pXMFk2\n" +
+        "cWtMRmVDNjVjMktzOHhRdWZvCi0tLSBWNUpzL2g0SnBENFhPR1BqMHhZMHpqRGtZ\n" +
+        "Y2FzUjBSUUNkSTFGMHRGbVQ0CkTlx8fbyyH+q2i65DGOJhYRgWIWsxGVE4yTJtrD\n" +
+        "NdqmZSVrmhIAfkALCQi1PQw3+N88xe+5zjHy4Lpxi8uN2GMlud6x49yzcYI=\n" +
+        "-----END AGE ENCRYPTED FILE-----"
     const plaintext = await timelockDecrypt(payloadFromGoImpl, mockClient)
 
-    expect(plaintext).to.equal("Hello World")
+    expect(plaintext).to.equal("why are you reading this secret message??")
 })
